# docs/swagger.yaml
basePath: /
...
host: localhost:3000
info:
  contact:
    email: support@aygit.com
    name: API Destek
    url: http://www.aygit.com/destek
  description: Bu servis, NES Özel Entegratör üzerinden e-fatura, e-arşiv ve ürün
    yönetim süreçlerini yönetir.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Aygıt Muhasebe Entegrasyonu API
  version: "1.0"
paths:
...
  /api/v1/{type}/tags:
    get:
      consumes:
      - application/json
      description: Etiketleri listeler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Tags
      tags:
      - Tags
    post:
      consumes:
      - application/json
      description: Etiket ekler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Tag
      tags:
      - Tags
  /api/v1/{type}/tags/{id}:
    delete:
      consumes:
      - application/json
      description: Etiket siler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete Tag
      tags:
      - Tags
    get:
      consumes:
      - application/json
      description: Sorgulanan etiketi getirir
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Tag By ID
      tags:
      - Tags
    put:
      consumes:
      - application/json
      description: Etiket günceller
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update Tag
      tags:
      - Tags
  /api/v1/{type}/uploads/document:
    post:
      consumes:
      - application/json
      description: Yeni bir E-MM veya E-SMM makbuz belgesi yükler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Belge verisi
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Upload Document (Belge Yükleme)
      tags:
      - Vouchers
  /api/v1/{type}/uploads/document/{uuid}:
    put:
      consumes:
      - multipart/form-data
      description: Belge günceller
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Document File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update Document Upload
      tags:
      - Uploads
  /api/v1/{type}/uploads/draft/create/{id}:
    post:
      consumes:
      - multipart/form-data
      description: Taslak Belge yükler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: Document File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Create Draft Upload
      tags:
      - Uploads
  /api/v1/{type}/uploads/draft/send:
    post:
      consumes:
      - application/json
      description: Taslak halindeki makbuzu resmileştirir
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Taslak UUID listesi
        in: body
        name: payload
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Send Draft (Taslak Gönderme)
      tags:
      - Vouchers
  /api/v1/{type}/vouchers:
    get:
      consumes:
      - application/json
      description: Taslak veya onaylı makbuzları listeler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Makbuz listesi
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Vouchers (Listeleme)
      tags:
      - Vouchers
...
  /api/v1/{type}/vouchers/drafts:
    delete:
      consumes:
      - application/json
      description: Taslak belgeleri silmek için bu uç kullanılablir
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      - description: Data
        in: body
        name: payload
        required: true
        schema: {}
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete Drafts
      tags:
      - Vouchers
    get:
      description: Taslak makbuzları listeler
      parameters:
      - description: 'Voucher Type: emm or esmm'
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Draft Vouchers (Taslak Makbuzları Listeleme)
      tags:
      - Vouchers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"